Use `user` as a string instead of `user_id` as an integer
To best support the widest range of unique user identifiers, switch to a
string. We don't care what it is, we just forward the value into the
`--user` parameter so no reason to restrict it to just numbers. Also,
everywhere else we just use `user` so it's more consistent internally as
well.